[Pack][Pin Counting] Move pin counting logic into ClusterPinCounter.
