+++ vpr/src/pack/cluster_pin_counter.cpp
@@ -6,9 +6,15 @@
 #include "cluster_pin_counter.h"
 
 #include <algorithm>
+#include <unordered_set>
 
+#include "atom_netlist.h"
+#include "atom_pb_bimap.h"
+#include "globals.h"
 #include "physical_types.h"
+#include "vpr_context.h"
 #include "vpr_types.h"
+#include "vpr_utils.h"
 #include "vtr_assert.h"
 
 void ClusterPinCounter::allocate_pb_state(const t_pb* pb) {
@@ -171,5 +177,287 @@ size_t ClusterPinCounter::committed_output_size(const t_pb* pb, size_t class_id)
     auto it = per_pb_state_.find(pb);
     VTR_ASSERT(it != per_pb_state_.end());
     return it->second.committed_output_pin_class_nets[class_id].size();
+}
+
+/*
+ * @brief Checks if the sinks of the given net are reachable from the driver
+ *        pb gpin.
+ */
+static int net_sinks_reachable_in_cluster(const t_pb_graph_pin* driver_pb_gpin, const int depth, const AtomNetId net_id, const AtomPBBimap& atom_to_pb) {
+    const AtomContext& atom_ctx = g_vpr_ctx.atom();
+
+    //Record the sink pb graph pins we are looking for
+    std::unordered_set<const t_pb_graph_pin*> sink_pb_gpins;
+    for (const AtomPinId pin_id : atom_ctx.netlist().net_sinks(net_id)) {
+        const t_pb_graph_pin* sink_pb_gpin = find_pb_graph_pin(atom_ctx.netlist(), atom_to_pb, pin_id);
+        VTR_ASSERT(sink_pb_gpin);
+
+        sink_pb_gpins.insert(sink_pb_gpin);
+    }
+
+    //Count how many sink pins are reachable
+    size_t num_reachable_sinks = 0;
+    for (int i_prim_pin = 0; i_prim_pin < driver_pb_gpin->num_connectable_primitive_input_pins[depth]; ++i_prim_pin) {
+        const t_pb_graph_pin* reachable_pb_gpin = driver_pb_gpin->list_of_connectable_input_pin_ptrs[depth][i_prim_pin];
+
+        if (sink_pb_gpins.count(reachable_pb_gpin)) {
+            ++num_reachable_sinks;
+            if (num_reachable_sinks == atom_ctx.netlist().net_sinks(net_id).size()) {
+                return true;
+            }
+        }
+    }
+
+    return false;
+}
+
+/**
+ * @brief Returns the pb_graph_pin of the atom pin defined by the driver_pin_id in the driver_pb
+ */
+static t_pb_graph_pin* get_driver_pb_graph_pin(const t_pb* driver_pb, const AtomPinId driver_pin_id) {
+    const AtomNetlist& atom_netlist = g_vpr_ctx.atom().netlist();
+
+    const auto driver_pb_type = driver_pb->pb_graph_node->pb_type;
+    int output_port = 0;
+    // find the port of the pin driving the net as well as the port model
+    auto driver_port_id = atom_netlist.pin_port(driver_pin_id);
+    auto driver_model_port = atom_netlist.port_model(driver_port_id);
+    // find the port id of the port containing the driving pin in the driver_pb_type
+    for (int i = 0; i < driver_pb_type->num_ports; i++) {
+        auto& prim_port = driver_pb_type->ports[i];
+        if (prim_port.type == OUT_PORT) {
+            if (prim_port.model_port == driver_model_port) {
+                // get the output pb_graph_pin driving this input net
+                return &(driver_pb->pb_graph_node->output_pins[output_port][atom_netlist.pin_port_bit(driver_pin_id)]);
+            }
+            output_port++;
+        }
+    }
+    // the pin should be found
+    VTR_ASSERT(false);
+    return nullptr;
+}
+
+void ClusterPinCounter::compute_and_mark_lookahead_pins_used_for_pin(
+    const t_pb_graph_pin* pb_graph_pin,
+    const t_pb* primitive_pb,
+    AtomNetId net_id,
+    const vtr::vector_map<AtomBlockId, LegalizationClusterId>& atom_cluster,
+    const AtomPBBimap& atom_to_pb) {
+    const AtomContext& atom_ctx = g_vpr_ctx.atom();
+
+    // starting from the parent pb of the input primitive go up in the hierarchy till the root block
+    for (auto cur_pb = primitive_pb->parent_pb; cur_pb; cur_pb = cur_pb->parent_pb) {
+        const auto depth = cur_pb->pb_graph_node->pb_type->depth;
+        const auto pin_class = pb_graph_pin->parent_pin_class[depth];
+        VTR_ASSERT(pin_class != UNDEFINED);
+
+        const auto driver_blk_id = atom_ctx.netlist().net_driver_block(net_id);
+
+        // if this primitive pin is an input pin
+        if (pb_graph_pin->port->type == IN_PORT) {
+            /* find location of net driver if exist in clb, NULL otherwise */
+            // find the driver of the input net connected to the pin being studied
+            const auto driver_pin_id = atom_ctx.netlist().net_driver(net_id);
+            // find the id of the atom occupying the input primitive_pb
+            const auto prim_blk_id = atom_to_pb.pb_atom(primitive_pb);
+            // find the pb block occupied by the driving atom
+            const auto driver_pb = atom_to_pb.atom_pb(driver_blk_id);
+            // pb_graph_pin driving net_id in the driver pb block
+            t_pb_graph_pin* output_pb_graph_pin = nullptr;
+            // if the driver block is in the same clb as the input primitive block
+            LegalizationClusterId driver_cluster_id = atom_cluster[driver_blk_id];
+            LegalizationClusterId prim_cluster_id = atom_cluster[prim_blk_id];
+            if (driver_cluster_id == prim_cluster_id) {
+                // get pb_graph_pin driving the given net
+                output_pb_graph_pin = get_driver_pb_graph_pin(driver_pb, driver_pin_id);
+            }
+
+            bool is_reachable = false;
+
+            // if the driver pin is within the cluster
+            if (output_pb_graph_pin) {
+                // find if the driver pin can reach the input pin of the primitive or not
+                const t_pb* check_pb = driver_pb;
+                while (check_pb && check_pb != cur_pb) {
+                    check_pb = check_pb->parent_pb;
+                }
+                if (check_pb) {
+                    for (int i = 0; i < output_pb_graph_pin->num_connectable_primitive_input_pins[depth]; i++) {
+                        if (pb_graph_pin == output_pb_graph_pin->list_of_connectable_input_pin_ptrs[depth][i]) {
+                            is_reachable = true;
+                            break;
+                        }
+                    }
+                }
+            }
+
+            // Must use an input pin to connect the driver to the input pin of the given primitive, either the
+            // driver atom is not contained in the cluster or is contained but cannot reach the primitive pin
+            if (!is_reachable) {
+                mark_lookahead_input(cur_pb, pin_class, net_id);
+            }
+        } else {
+            VTR_ASSERT(pb_graph_pin->port->type == OUT_PORT);
+            /*
+             * Determine if this net (which is driven from within this cluster) leaves this cluster
+             * (and hence uses an output pin).
+             */
+
+            bool net_exits_cluster = true;
+            int num_net_sinks = static_cast<int>(atom_ctx.netlist().net_sinks(net_id).size());
+
+            if (pb_graph_pin->num_connectable_primitive_input_pins[depth] >= num_net_sinks) {
+                //It is possible the net is completely absorbed in the cluster,
+                //since this pin could (potentially) drive all the net's sinks
+
+                /* Important: This runtime penalty looks a lot scarier than it really is.
+                 * For high fan-out nets, I at most look at the number of pins within the
+                 * cluster which limits runtime.
+                 *
+                 * DO NOT REMOVE THIS INITIAL FILTER WITHOUT CAREFUL ANALYSIS ON RUNTIME!!!
+                 *
+                 * Key Observation:
+                 * For LUT-based designs it is impossible for the average fanout to exceed
+                 * the number of LUT inputs so it's usually around 4-5 (pigeon-hole argument,
+                 * if the average fanout is greater than the number of LUT inputs, where do
+                 * the extra connections go?  Therefore, average fanout must be capped to a
+                 * small constant where the constant is equal to the number of LUT inputs).
+                 * The real danger to runtime is when the number of sinks of a net gets doubled
+                 */
+
+                //Check if all the net sinks are, in fact, inside this cluster
+                bool all_sinks_in_cur_cluster = true;
+                LegalizationClusterId driver_cluster = atom_cluster[driver_blk_id];
+                for (auto pin_id : atom_ctx.netlist().net_sinks(net_id)) {
+                    auto sink_blk_id = atom_ctx.netlist().pin_block(pin_id);
+                    if (atom_cluster[sink_blk_id] != driver_cluster) {
+                        all_sinks_in_cur_cluster = false;
+                        break;
+                    }
+                }
+
+                if (all_sinks_in_cur_cluster) {
+                    //All the sinks are part of this cluster, so the net may be fully absorbed.
+                    //
+                    //Verify this, by counting the number of net sinks reachable from the driver pin.
+                    //If the count equals the number of net sinks then the net is fully absorbed and
+                    //the net does not exit the cluster
+                    /* TODO: I should cache the absorbed outputs, once net is absorbed,
+                     *       net is forever absorbed, no point in rechecking every time */
+                    if (net_sinks_reachable_in_cluster(pb_graph_pin, depth, net_id, atom_to_pb)) {
+                        //All the sinks are reachable inside the cluster
+                        net_exits_cluster = false;
+                    }
+                }
+            }
+
+            if (net_exits_cluster) {
+                /* This output must exit this cluster */
+                mark_lookahead_output(cur_pb, pin_class, net_id);
+            }
+        }
+    }
+}
+
+void ClusterPinCounter::compute_and_mark_lookahead_pins_used(
+    AtomBlockId blk_id,
+    const vtr::vector_map<AtomBlockId, LegalizationClusterId>& atom_cluster,
+    const AtomPBBimap& atom_to_pb) {
+    const AtomNetlist& atom_netlist = g_vpr_ctx.atom().netlist();
+
+    const t_pb* cur_pb = atom_to_pb.atom_pb(blk_id);
+    VTR_ASSERT(cur_pb != nullptr);
+
+    /* Walk through inputs, outputs, and clocks marking pins off of the same class */
+    for (auto pin_id : atom_netlist.block_pins(blk_id)) {
+        auto net_id = atom_netlist.pin_net(pin_id);
+
+        const t_pb_graph_pin* pb_graph_pin = find_pb_graph_pin(atom_netlist, atom_to_pb, pin_id);
+        compute_and_mark_lookahead_pins_used_for_pin(pb_graph_pin, cur_pb, net_id, atom_cluster, atom_to_pb);
+    }
+}
+
+void ClusterPinCounter::try_update_lookahead_pins_used(
+    const std::vector<PackMoleculeId>& molecules,
+    const Prepacker& prepacker,
+    const vtr::vector_map<AtomBlockId, LegalizationClusterId>& atom_cluster,
+    const AtomPBBimap& atom_to_pb) {
+    for (PackMoleculeId molecule_id : molecules) {
+        const t_pack_molecule& molecule = prepacker.get_molecule(molecule_id);
+        for (AtomBlockId blk_id : molecule.atom_block_ids) {
+            if (!blk_id.is_valid()) {
+                continue;
+            }
+
+            const t_pb* primitive_pb = atom_to_pb.atom_pb(blk_id);
+            VTR_ASSERT_SAFE(primitive_pb != nullptr);
+            VTR_ASSERT_SAFE(primitive_pb->pb_graph_node->pb_type->is_primitive());
+
+            VTR_ASSERT(primitive_pb->pb_graph_node->pb_type->blif_model != nullptr);
+            compute_and_mark_lookahead_pins_used(blk_id, atom_cluster, atom_to_pb);
+        }
+    }
+}
+
+bool ClusterPinCounter::check_lookahead_pins_used(t_pb* cur_pb, t_ext_pin_util max_external_pin_util) {
+    const t_pb_type* pb_type = cur_pb->pb_graph_node->pb_type;
+
+    if (!pb_type->is_primitive() && cur_pb->name) {
+        for (size_t class_id = 0; class_id < cur_pb->pb_graph_node->input_pin_class_sizes.size(); class_id++) {
+            size_t class_size = cur_pb->pb_graph_node->input_pin_class_sizes[class_id];
+
+            if (cur_pb->is_root()) {
+                // Scale the class size by the maximum external pin utilization factor
+                // Use ceil to avoid classes of size 1 from being scaled to zero
+                class_size = std::ceil(max_external_pin_util.input_pin_util * class_size);
+                // if the number of pins already used is larger than class size, then the number of
+                // cluster inputs already used should be our constraint. Why is this needed? This is
+                // needed since when packing the seed block the maximum external pin utilization is
+                // used as 1.0 allowing molecules that are using up to all the cluster inputs to be
+                // packed legally. Therefore, if the seed block is already using more inputs than
+                // the allowed maximum utilization, this should become the new maximum pin utilization.
+                class_size = std::max<size_t>(class_size, committed_input_size(cur_pb, class_id));
+            }
+
+            if (lookahead_input_size(cur_pb, class_id) > class_size) {
+                return false;
+            }
+        }
+
+        for (size_t class_id = 0; class_id < cur_pb->pb_graph_node->output_pin_class_sizes.size(); class_id++) {
+            size_t class_size = cur_pb->pb_graph_node->output_pin_class_sizes[class_id];
+
+            if (cur_pb->is_root()) {
+                // Scale the class size by the maximum external pin utilization factor
+                // Use ceil to avoid classes of size 1 from being scaled to zero
+                class_size = std::ceil(max_external_pin_util.output_pin_util * class_size);
+                // if the number of pins already used is larger than class size, then the number of
+                // cluster outputs already used should be our constraint. Why is this needed? This is
+                // needed since when packing the seed block the maximum external pin utilization is
+                // used as 1.0 allowing molecules that are using up to all the cluster inputs to be
+                // packed legally. Therefore, if the seed block is already using more inputs than
+                // the allowed maximum utilization, this should become the new maximum pin utilization.
+                class_size = std::max<size_t>(class_size, committed_output_size(cur_pb, class_id));
+            }
+
+            if (lookahead_output_size(cur_pb, class_id) > class_size) {
+                return false;
+            }
+        }
+
+        if (cur_pb->child_pbs) {
+            for (int i = 0; i < pb_type->modes[cur_pb->mode].num_pb_type_children; i++) {
+                if (cur_pb->child_pbs[i]) {
+                    for (int j = 0; j < pb_type->modes[cur_pb->mode].pb_type_children[i].num_pb; j++) {
+                        if (!check_lookahead_pins_used(&cur_pb->child_pbs[i][j], max_external_pin_util))
+                            return false;
+                    }
+                }
+            }
+        }
+    }
+
+    return true;
 }
